+++ icaruscode/TPC/NuGraph/testinference_slice_icarus.fcl
@@ -1,5 +1,5 @@
 ###
-### Test fcl file for NuGraph inference. For each cryostat it runs two producers (IcarusNuSliceHitsProducer and NuGraphInference) plus one analyzer (NuGraphAnalyzer).
+### Test fcl file for NuGraph inference. For each cryostat it runs two producers (ICARUSNuSliceHitsProducer and NuGraphInference) plus one analyzer (NuGraphAnalyzer).
 ### Author: G. Cerati (FNAL)
 ###
 
@@ -89,8 +89,8 @@ physics:
   end_paths:          [ ana, streamROOT ]
 }
 
-#physics.producers.nuslhitsCryoE.module_type: "IcarusTrueNuSliceHitsProducer"
-#physics.producers.nuslhitsCryoW.module_type: "IcarusTrueNuSliceHitsProducer"
+#physics.producers.nuslhitsCryoE.module_type: "ICARUSTrueNuSliceHitsProducer"
+#physics.producers.nuslhitsCryoW.module_type: "ICARUSTrueNuSliceHitsProducer"
 
 #physics.producers.NuGraphCryoE.debug: true
 #physics.producers.NuGraphCryoW.debug: true
@@ -106,7 +106,7 @@ services.BackTrackerService.BackTracker.SimChannelModuleLabel: "daq:simpleSC"
 # then after the job you can kill the server process with:
 # > killall -9 /cvmfs/oasis.opensciencegrid.org/mis/apptainer/1.3.2/x86_64/libexec/apptainer/libexec/starter
 #NuGraphCryoE: @local::ApptainerNuGraphNuSonicTriton
-#NuGraphCryoE.LoaderTool.tool_type: "IcarusNuGraphLoader"
+#NuGraphCryoE.LoaderTool.tool_type: "ICARUSNuGraphLoader"
 #NuGraphCryoE.LoaderTool.hitInput: "nuslhitsCryoE"
 #NuGraphCryoE.LoaderTool.spsInput: "nuslhitsCryoE"
 #NuGraphCryoE.DecoderTools.SemanticDecoderTool.hitInput: "nuslhitsCryoE"
